--- src/Pages/ContactForm/ContactForm.js
@@ -1,5 +1,0 @@
-const ContactFormPage = () => {
-  return <h1>I am a ContactForm Page Component</h1>;
-};
-
-export default ContactFormPage;
